backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
werkzeug
PyJWT
reportlab
gunicorn
psycopg2-binary
